feat(www): make dodont available in patterns
